+++ CHANGELOG.md
@@ -10,6 +10,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Fixed
 
 -   Wireless Crafting Grid still requiring energy even if Refined Storage is configured not to use energy.
+-   You can no longer change the energy capacity to be zero.
 
 ## [1.0.6] - 2025-09-12